노트 제목 수정 기능 추가 (#6)
* [25.05.13] 작업내역
1. 가림판 선택화면 수정 (미리보기 추가)
2. 가림판 편집 기능 추가

* [25.05.14] 가림판 선택 화면 css 수정

* [25.05.15] 배경 이미지 파일 삭제 로직 수정

* [25.05.16] 가림판 삭제 기능 추가

* [25.05.16] 가림판 삭제 시 이미지 파일도 함께 삭제되도록 수정

* [25.05.18] 가림판 이미지 파일명 저장 규칙 수정 및  split페이지 인덱스 css 수정

* [25.05.18] 북마크 바로가기 클릭 이벤트 수정

* [25.05.18] 사용중인 가림판 수정 시 설정이 바로 적용되도록 수정

* [25.05.18] DB charset 설정 수정

* [25.05.18] 가림판 등록 갯수 제한 추가, 가림판 이미지 src값 오류 수정

* [25.05.18] 메인화면 노트 제목 수정 기능 추가 및 말줄임표 처리

---------

+++ public/js/cover/list.js
@@ -30,7 +30,6 @@ async function chageCoverSelect() {
         else {
             setCoverOpt(data); // 미리보기 화면 적용
         }
-
     } catch (error) {
         console.error('예외 발생:', error);
         alert('네트워크 오류가 발생했습니다.');
@@ -66,8 +65,6 @@ async function setDefaultCover() {
 
 // 가림판 설정 적용
 function setCoverOpt(data) {
-    console.log(data);
-
     const coverPreview = document.getElementById("coverPreview");//가림판 미리보기 화면
     const coverImage = document.getElementById("previewImage");
     const decorationText = document.getElementById('decorationText');//꾸밈 문구
@@ -93,7 +90,15 @@ async function goCoverEdit(btnName) {
     if (btnName == "update") {
         cover_id = document.getElementsByName("cover_id")[0].value;
         if (cover_id == '-1') {
-            alert("기본 커버는 편집이 불가능 합니다.");
+            alert("기본 가림판은 편집이 불가능 합니다.");
+            return;
+        }
+    }
+    else {
+        // 가림판 갯수가 6개 이상이면 등록 불가
+        const coverCount = document.getElementById('coverSelectbox').options.length
+        if (coverCount > 5) {
+            alert("최대 5개 까지만 가림판 등록이 가능합니다.");
             return;
         }
     }
@@ -109,24 +114,26 @@ async function deleteCover() {
             return;
         }
 
-        const response = await fetch('/cover/delete', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({ cover_id: cover_id })
-        });
+        if (confirm("정말 삭제하시겠습니까?")) {
+            const response = await fetch('/cover/delete', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({ cover_id: cover_id })
+            });
 
-        const data = await response.json();
-        if (!response.ok) {
-            alert(data.message);
-        } else {
-            alert(data.message);
+            const data = await response.json();
+            if (!response.ok) {
+                alert(data.message);
+            } else {
+                alert(data.message);
 
-            const coverImage = document.getElementById("previewImage").src;
-            const filename = coverImage.split("/uploads/")[1];
-            // 서버에 저장된 이미지 삭제 
-            if (filename) await deleteImage(filename);
+                const coverImage = document.getElementById("previewImage").getAttribute('src');
+                const filename = coverImage.split("/uploads/")[1];
+                // 서버에 저장된 이미지 삭제 
+                if (filename) await deleteImage(filename);
+            }
         }
     }
     catch (error) {
